initial moving of all files into new structure
