--- src/gui/popup_windows/frame_range_dialog.py
@@ -1,68 +1,0 @@
-from PyQt6.QtCore import Qt
-from PyQt6.QtWidgets import QDialog, QLineEdit, QDialogButtonBox, QFormLayout
-
-
-class FrameRangeDialog(QDialog):
-    def __init__(self, main_window, step: bool = False):
-        super().__init__(main_window)
-        self.main_window = main_window
-        self.lower_limit = QLineEdit(self)
-        self.lower_limit.setText('1')
-        self.upper_limit = QLineEdit(self)
-        self.upper_limit.setText(str(main_window.runtime_data.images.shape[0]))
-        if step:
-            self.step_layout = QLineEdit(self)
-            self.step_layout.setText('1')
-
-        buttonBox = QDialogButtonBox(QDialogButtonBox.StandardButton.Ok | QDialogButtonBox.StandardButton.Cancel, self)
-
-        layout = QFormLayout(self)
-        layout.addRow('Lower limit', self.lower_limit)
-        layout.addRow('Upper limit', self.upper_limit)
-        if step:
-            layout.addRow('Step (mm)', self.step_layout)
-        layout.addWidget(buttonBox)
-
-        buttonBox.accepted.connect(self.accept)
-        buttonBox.rejected.connect(self.reject)
-
-    def getInputs(self):
-        lower_limit = int(self.lower_limit.text()) - 1
-        lower_limit = max(0, lower_limit)
-        upper_limit = int(self.upper_limit.text())
-        upper_limit = min(self.main_window.runtime_data.images.shape[0], upper_limit)
-
-        if lower_limit >= upper_limit:
-            lower_limit, upper_limit = upper_limit, lower_limit
-        return lower_limit, upper_limit
-
-    def getStep(self):
-        return float(self.step_layout.text())
-
-
-class StartFramesDialog(QDialog):
-    def __init__(self, main_window, label1='First diastolic frame', label2='First systolic frame'):
-        super().__init__(main_window)
-        self.main_window = main_window
-
-        self.diastolic_start = QLineEdit(self)
-        self.systolic_start = QLineEdit(self)
-
-        buttonBox = QDialogButtonBox(QDialogButtonBox.StandardButton.Ok | QDialogButtonBox.StandardButton.Cancel, self)
-
-        layout = QFormLayout(self)
-        layout.addRow(label1, self.diastolic_start)
-        layout.addRow(label2, self.systolic_start)
-        layout.addWidget(buttonBox)
-
-        buttonBox.accepted.connect(self.accept)
-        buttonBox.rejected.connect(self.reject)
-
-        # Set non-modal mode
-        self.setWindowModality(Qt.WindowModality.NonModal)
-
-    def getInputs(self):
-        # Retrieve and return input values
-        diastolic = int(self.diastolic_start.text()) - 1
-        systolic = int(self.systolic_start.text()) - 1
-        return diastolic, systolic
